益田さんが作成したapi呼び出しの方法に対応
